Crate azalea_inventory

Modules

components

Representations of various inventory data structures in Minecraft.

item

operations

Structs

DataComponentPatch

ItemSlotData

An item in an inventory, with a count and NBT. Usually you want ItemSlot or [azalea_registry::Item] instead.

Player

SlotList

A fixed-size list of ItemSlots.

Enums

AnvilMenuLocation

BeaconMenuLocation

BlastFurnaceMenuLocation

BrewingStandMenuLocation

CartographyTableMenuLocation

Crafter3x3MenuLocation

CraftingMenuLocation

EnchantmentMenuLocation

FurnaceMenuLocation

Generic3x3MenuLocation

Generic9x1MenuLocation

Generic9x2MenuLocation

Generic9x3MenuLocation

Generic9x4MenuLocation

Generic9x5MenuLocation

Generic9x6MenuLocation

GrindstoneMenuLocation

HopperMenuLocation

ItemSlot

Either an item in an inventory or nothing.

LecternMenuLocation

LoomMenuLocation

Menu

A menu, which is a fixed collection of slots.

MenuLocation

MerchantMenuLocation

PlayerMenuLocation

ShulkerBoxMenuLocation

SmithingMenuLocation

SmokerMenuLocation

StonecutterMenuLocation