azalea_inventory
Enum ItemSlot

pub enum ItemSlot {
    Empty,
    Present(ItemSlotData),
}

Either an item in an inventory or nothing.

Variants

Empty

Present(ItemSlotData)

Implementations

impl ItemSlot

pub fn is_empty(&self) -> bool

Check if the slot is ItemSlot::Empty, if the count is <= 0, or if the item is air.

This is the opposite of ItemSlot::is_present.

pub fn is_present(&self) -> bool

Check if the slot is not ItemSlot::Empty, if the count is > 0, and if the item is not air.

This is the opposite of ItemSlot::is_empty.

pub fn count(&self) -> i32

Return the amount of the item in the slot, or 0 if the slot is empty.

Note that itโs possible for the count to be zero or negative when the slot is present.

pub fn split(&mut self, count: u32) -> ItemSlot

Remove count items from this slot, returning the removed items.

pub fn kind(&self) -> Item

Get the kind of the item in this slot, or [azalea_registry::Item::Air]

pub fn update_empty(&mut self)

Update whether this slot is empty, based on the count.

pub fn as_present(&self) -> Option<&ItemSlotData>

Convert this slot into an ItemSlotData, if itโs present.