azalea_inventory::item
Trait MaxStackSizeExt

pub trait MaxStackSizeExt {
    // Required method
    fn max_stack_size(&self) -> i32;

    // Provided method
    fn stackable(&self) -> bool { ... }
}

Required Methods

fn max_stack_size(&self) -> i32

Get the maximum stack size for this item.

This is a signed integer to be consistent with the count field of ItemSlotData.

Provided Methods

fn stackable(&self) -> bool

Whether this item can be stacked with other items.

This is equivalent to self.max_stack_size() > 1.